requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-crum==0.7.7
django-easy-pdf==0.1.1
django-import-export==2.4.0
django-userforeignkey==0.3.0
djangorestframework==3.12.1
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
psycopg2==2.8.6
PyJWT==1.7.1
pylint==2.6.0
PyPDF2==1.26.0
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.46
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.2
twilio==6.45.4
typed-ast==1.4.1
urllib3==1.25.10
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.12.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
